Moved components to different folder Added eslint configuration Added linting scripts
